Relocate calendarOps into domains/calendar/service (pure move)
git mv lib/calendarOps.ts -> domains/calendar/service.ts (+ its test), the
Google-write orchestration ops shared by the calendar actions and the
assistant's applyProposal. Import depth fixed; the three importers (calendar,
assistant, assistantTools) repointed. First step of the calendar-domain reorg,
kept separate so it reads as a clean relocation. No behavior change — git
detects the rename, service tests (20) + full suite green, api.d.ts diff only
swaps the module-map path.
